Rename kilnUtils.ts -> KilnApi.ts
