+++ src/plugins/stake-plugins/generic/policyAdapters/CardanoKilnAdaptor.ts
@@ -7,7 +7,7 @@ import { infoServerData } from '../../../../util/network'
 import { AssetId, ChangeQuote, PositionAllocation, QuoteAllocation, StakePosition } from '../../types'
 import { asInfoServerResponse } from '../../util/internalTypes'
 import { StakePolicyConfig } from '../types'
-import { KilnError, makeKilnApi } from '../util/kilnUtils'
+import { KilnError, makeKilnApi } from '../util/KilnApi'
 import { StakePolicyAdapter } from './types'
 
 const MIN_STAKE_LOVELACE_AMOUNT = '2300000'
